fix: agent callback with step infomation (#774)
