+++ packages/core/src/callbacks/CallbackManager.ts
@@ -1,7 +1,7 @@
 import type { Anthropic } from "@anthropic-ai/sdk";
 import { CustomEvent } from "@llamaindex/env";
 import type { NodeWithScore } from "../Node.js";
-import type { AgentEndEvent, AgentStartEvent } from "../agent/type.js";
+import type { AgentEndEvent, AgentStartEvent } from "../agent/types.js";
 import {
   EventCaller,
   getEventCaller,
