+++ packages/core/src/index.ts
@@ -1,5 +1,10 @@
 export * from "./index.edge.js";
 export * from "./readers/index.js";
 export * from "./storage/index.js";
+// Exports modules that doesn't support non-node.js runtime
 // Ollama is only compatible with the Node.js runtime
+export {
+  HuggingFaceEmbedding,
+  HuggingFaceEmbeddingModelType,
+} from "./embeddings/HuggingFaceEmbedding.js";
 export { Ollama, type OllamaParams } from "./llm/ollama.js";
