+++ packages/core/src/internal/type.ts
@@ -1,5 +1,5 @@
-import { CustomEvent } from "@llamaindex/env";
+import type { CustomEvent } from "@llamaindex/env";
 
 export type BaseEvent<Payload extends Record<string, unknown>> = CustomEvent<{
-  payload: Payload;
+  payload: Readonly<Payload>;
 }>;
